% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMeritPlot.R
\name{customMeritPlot}
\alias{customMeritPlot}
\title{Plot merit order curve}
\usage{
customMeritPlot(data, prix, demand)
}
\arguments{
\item{data}{A dataframe containing the data to be plotted. It should contain columns 'position', 'MC' (Marginal costs), 'Capacity' and 'Type'}

\item{prix}{The equilibrium price to be plotted on the y-axis}

\item{demand}{The demand to be plotted on the x-axis}
}
\value{
A plotly object of the merit order curve
}
\description{
Plot merit order curve
}
