% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customPlot.R
\name{customPlot}
\alias{customPlot}
\title{Render a line plot}
\usage{
customPlot(data, serie, lim)
}
\arguments{
\item{data}{A dataframe containing the historical data for the specified serie}

\item{serie}{The name of the serie for which the plot will be rendered}

\item{lim}{A value used as a threshold for the plot, a line will be drawn on this value}
}
\value{
A plotly object containing the line plot of the specified serie with additional lines for the low and high deviation scenarios
}
\description{
Render a line plot
}
