% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customTable.R
\name{customTable}
\alias{customTable}
\title{customTable - Renders a table with options for color and formatting}
\usage{
customTable(data, serie, lim)
}
\arguments{
\item{data}{A dataframe containing the historical data for the specified serie}

\item{serie}{The name of the serie to be displayed in the table}

\item{lim}{The limit used to determine the color of the rows}
}
\value{
A rendered datatable with options for color, formatting and row grouping
}
\description{
customTable - Renders a table with options for color and formatting
}
