% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit models}
\usage{
fit(train, test, h, serie)
}
\arguments{
\item{train}{A dataframe containing the training dataset.}

\item{test}{A dataframe containing the test dataset.}

\item{h}{The number of steps ahead to forecast.}

\item{serie}{The name of the column containing the time series data.}
}
\value{
A dataframe containing the forecasted values and the actual values.
}
\description{
Fit models
}
