% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{scenario}
\alias{scenario}
\title{Compute low and high deviation scenarios}
\usage{
scenario(data, serie, last_date, low, high, all)
}
\arguments{
\item{data}{A dataframe containing the historical data for the specified serie}

\item{serie}{The name of the serie for which the low and high deviation scenarios will be computed}

\item{last_date}{The date of the last observation used to compute the scenarios}

\item{low}{The low deviation percentage (default: 5)}

\item{high}{The high deviation percentage (default: 5)}

\item{all}{If TRUE, return all columns of the dataframe, if FALSE, return only the Date, serie, serie_low and serie_high columns (default: TRUE)}
}
\value{
A dataframe containing the historical data with additional columns for the expected values, low deviation and high deviation scenarios
}
\description{
Compute low and high deviation scenarios
}
